Map mapping characters remappable.
Unfortunately the current revision of the autoclose.vim script does not
work with g:AutoClosePairs when it contains '|'.

Specifically, the vim [re]map and execute commands cannot take the '|'
symbol as an argument as it also acts as a command separator (like ';'
is in C).

Now there is a new internal map that can be used to inform those special
maps.

The whole solution was proposed and coded by Benjamin Howell.

Thanks Benjamin!