Add an option to not use the "preserve dot register" feature.
If you set g:AutoClosePreservDotReg to 0 on your vimrc, AutoClose will not use the completion buffer and will not map the flushing keys (Esc, Left, Right, Up, Down, MouseLeftButton and MouseRightButton).

This option is for those who are getting problems with the arrow keys on vim inside a terminal. Without this, the Esc key is always mapped, therefore when a terminal sends the sequence ^[OA to vim (which is an up arrow), vim will receive the <Esc> char (^[) and trigger its map, which makes the arrow keys insert a new line with a upper letter on it in the sequence.