Refactoring: dict defaults and popup menu fixes
This uses the method introduces for movementKeys before for all the
settings.

Also improve default pairs: add '<>', '``' and '«»'.

While the latter is not on regular keyboards, it is provided at least on
the Neo keyboard layout.

Also popup menu related fixes:

Set AutoClosePumvisible* to the key itself by default.

This fixes using the movement keys in a menu, e.g. with the following
text:
  foo
  fooo
  fo_

Typing "fo" opens the autocomplpop popup; using the cursor keys there
made the menu disappear and the movement carried out instead.

pumvisible() is handled depending on
the existence of a Pumvisible* setting.

Only Left, Right, PageUp and PageDown are handled as-is in the
completion menu. All other movement keys will leave the menu and need to
flush the buffer.