Make AutoClosePairs filetype-aware.
Add `reset` argument to DefineVariables which gets enabled on FileType
events and will reset all settings in the buffer namespace.
An additional (experimental) check looks at the filetype and will add
the '||' pair for &ft=ruby.

See
https://github.com/thenoseman/vim-autoclose/commit/128c289d8ab6b52b0ac723dfd25fcb9387fabd85#commitcomment-324879
for the idea.

Remove '<'/'>' map for typoscript and shell.

This is still experimental and needs some refactoring regarding the lookup.